Upgrade strategy.
